+++ Cargo.toml
@@ -8,7 +8,7 @@ build = "build.rs"
 [dependencies]
 anyhow = "1.0.102"
 cwtools_parser = { git = "https://github.com/MillenniumDawn/cwtools.git", rev = "cb7f8c8c95dd68fcbd06d37efb87bc49fc17fbef", package = "cwtools_parser" }
-cwtools-hoi4-config = { git = "https://github.com/Hilde-Research-Foundation/cwtools-hoi4-config-rs.git", rev = "c8bc2d825d066ec69e413ee9ace9c1ffc02103f3" }
+cwtools-hoi4-config = { git = "https://github.com/Hilde-Research-Foundation/cwtools-hoi4-config-rs.git", rev = "d40f0f65137e5e96c044a6d848653135d2b22fec" }
 cwtools_rules = { git = "https://github.com/MillenniumDawn/cwtools.git", rev = "cb7f8c8c95dd68fcbd06d37efb87bc49fc17fbef", package = "cwtools_rules" }
 cwtools_string_table = { git = "https://github.com/MillenniumDawn/cwtools.git", rev = "cb7f8c8c95dd68fcbd06d37efb87bc49fc17fbef", package = "cwtools_string_table" }
 cwtools_validation = { git = "https://github.com/MillenniumDawn/cwtools.git", rev = "cb7f8c8c95dd68fcbd06d37efb87bc49fc17fbef", package = "cwtools_validation" }
@@ -36,7 +36,8 @@ toml = "0.8.23"
 rustix = { version = "1.1.4", features = ["fs"] }
 
 [target.'cfg(windows)'.dependencies]
-windows-sys = { version = "0.61.2", features = ["Win32_Foundation", "Win32_Storage_FileSystem"] }
+atomicwrites = "=0.4.4"
+file-id = "=0.2.3"
 
 [profile.release]
 codegen-units = 1
